Move verilator to testing subpackage
